Add utilities to encode/decode data classes to/from collections
