+++ core/src/main/kotlin/com/hexagonkt/core/Helpers.kt
@@ -10,7 +10,6 @@ import java.net.Socket
 import java.util.*
 import java.util.concurrent.TimeUnit.SECONDS
 import kotlin.reflect.KClass
-import kotlin.reflect.KProperty1
 
 /**
  *  Disable heavy and optional checks in runtime. This flag can be enabled to get a small
@@ -238,148 +237,3 @@ fun Throwable.toText(prefix: String = ""): String =
             ""
         else
             "${eol}Caused by: " + (this.cause as Throwable).toText(prefix)
-
-// COLLECTIONS /////////////////////////////////////////////////////////////////////////////////////
-/**
- * [TODO](https://github.com/hexagonkt/hexagon/issues/271).
- *
- * @receiver .
- * @param count .
- * @return .
- */
-fun <Z> Collection<Z>.ensureSize(count: IntRange): Collection<Z> = this.apply {
-    if (size !in count) error("$size items while expecting only $count element")
-}
-
-/**
- * [TODO](https://github.com/hexagonkt/hexagon/issues/271).
- *
- * @receiver .
- * @param keys .
- * @return .
- */
-@Suppress("UNCHECKED_CAST")
-inline fun <reified T : Any> Map<*, *>.keys(vararg keys: Any): T? {
-
-    val mappedKeys = keys.map {
-        when (it) {
-            is KProperty1<*, *> -> it.name
-            else -> it
-        }
-    }
-
-    return mappedKeys
-        .dropLast(1)
-        .fold(this) { result, element ->
-            val r = result as Map<Any, Any>
-            when (val value = r[element]) {
-                is Map<*, *> -> value
-                is List<*> -> value.mapIndexed { ii, item -> ii to item }.toMap()
-                else -> emptyMap<Any, Any>()
-            }
-        }[mappedKeys.last()] as? T
-}
-
-/**
- * [TODO](https://github.com/hexagonkt/hexagon/issues/271).
- *
- * @receiver .
- * @param keys .
- * @return .
- */
-inline operator fun <reified T : Any> Map<*, *>.invoke(vararg keys: Any): T? =
-    keys(*keys)
-
-/**
- * [TODO](https://github.com/hexagonkt/hexagon/issues/271).
- *
- * @receiver .
- * @param name .
- * @return .
- */
-inline fun <reified T : Any> Map<*, *>.requireKeys(vararg name: Any): T =
-    this.keys(*name) ?: error("$name required key not found")
-
-/**
- * [TODO](https://github.com/hexagonkt/hexagon/issues/271).
- *
- * @param fields .
- * @return .
- */
-fun <T : Any> fieldsMapOf(vararg fields: Pair<KProperty1<T, *>, *>): Map<String, *> =
-    fields.associate { it.first.name to it.second }
-
-/**
- * [TODO](https://github.com/hexagonkt/hexagon/issues/271).
- *
- * @receiver .
- * @param name .
- * @return .
- */
-fun <K, V> Map<K, V>.require(name: K): V =
-    this[name] ?: error("$name required key not found")
-
-/**
- * [TODO](https://github.com/hexagonkt/hexagon/issues/271).
- *
- * @receiver .
- * @return .
- */
-fun <K, V> Map<K, V?>.filterEmpty(): Map<K, V> =
-    this.filterValues(::notEmpty).mapValues { (_, v) -> v ?: fail }
-
-/**
- * [TODO](https://github.com/hexagonkt/hexagon/issues/271).
- *
- * @receiver .
- * @return .
- */
-fun <V> List<V?>.filterEmpty(): List<V> =
-    this.filter(::notEmpty).map { it ?: fail }
-
-/**
- * [TODO](https://github.com/hexagonkt/hexagon/issues/271).
- *
- * @receiver .
- * @return .
- */
-fun Map<*, *>.filterEmptyRecursive(): Map<*, *> =
-    mapValues { (_, v) ->
-        when (v) {
-            is List<*> -> v.filterEmptyRecursive()
-            is Map<*, *> -> v.filterEmptyRecursive()
-            else -> v
-        }
-    }
-    .filterEmpty()
-
-/**
- * [TODO](https://github.com/hexagonkt/hexagon/issues/271).
- *
- * @receiver .
- * @return .
- */
-fun List<*>.filterEmptyRecursive(): List<*> =
-    map {
-        when (it) {
-            is List<*> -> it.filterEmptyRecursive()
-            is Map<*, *> -> it.filterEmptyRecursive()
-            else -> it
-        }
-    }
-    .filterEmpty()
-
-/**
- * [TODO](https://github.com/hexagonkt/hexagon/issues/271).
- *
- * @param value .
- * @return .
- */
-fun <V> notEmpty(value: V?): Boolean {
-    return when (value) {
-        null -> false
-        is List<*> -> value.isNotEmpty()
-        is Map<*, *> -> value.isNotEmpty()
-        else -> true
-    }
-}
